cleanup: factorize stack check
process() is where the cmds are called, that is the only place we could
check the stack has enough elements for the command, but as we use
__contains__ to check the command exist, we cannot replace the
dictionnary of commands by a tuple containing the length too.

So I went with a 2nd synced dict for the number of elements needed, init
is less readable, but it does avoid a lot of checks that are basically
duplicate code.